test: split format tests into format.test.ts
